changed currencies to decimals

+++ my_app/forms.py
@@ -22,22 +22,22 @@ class SignUpForm(UserCreationForm):
         fields = ('username', 'birth_date', 'password1', 'password2', 'location')
 
 class TopUpForm(ModelForm):
-    USD = forms.IntegerField(label='USD:')
-    EUR = forms.IntegerField(label='EUR:')
+    USD = forms.DecimalField(label='USD:')
+    EUR = forms.DecimalField(label='EUR:')
     class Meta:
         model = User
         fields = ('USD', 'EUR')
 
 class WithdrawForm(ModelForm):
-    USD = forms.IntegerField(label='USD:')
-    EUR = forms.IntegerField(label='EUR:')
+    USD = forms.DecimalField(label='USD:')
+    EUR = forms.DecimalField(label='EUR:')
     class Meta:
         model = User
         fields = ('USD', 'EUR')
 
 class TransferForm(ModelForm):
-    USD = forms.IntegerField(label='USD:')
-    EUR = forms.IntegerField(label='EUR:')
+    USD = forms.DecimalField(label='USD:')
+    EUR = forms.DecimalField(label='EUR:')
     username = forms.CharField(label='User:')
     class Meta:
         model = User
